Changes to MOAIActionMgr, MOAIInputMgr, graphics resource lifecycle

+++ src/moai-box2d/MOAIBox2DWorld.cpp
@@ -1130,7 +1130,7 @@ MOAIBox2DWorld::~MOAIBox2DWorld () {
 }
 
 //----------------------------------------------------------------//
-void MOAIBox2DWorld::OnUpdate ( float step ) {
+void MOAIBox2DWorld::OnUpdate ( double step ) {
 	
 	this->mLock = true;
 	this->mWorld->Step ( step, this->mVelocityIterations, this->mPositionIterations );
